Configure tdChai shim to import td correctly
Since `ember-cli-testdouble` now imports `testdouble` as a proper AMD
module, it no longer exposes the global that we were relying on. This
change tries import as a module first, falling back to the `window`
object instead.  It also tries to ensure that `ember-cli-testdouble`, if
installed, runs first, as suggested by @Turbo87.

Closes #28